nav bar unlogged, secu input, gloabl css

+++ controllers/articlemodify.js
@@ -11,6 +11,8 @@ const modifyArticle = (req, res) => {
     
     const form = formidable({keepExtensions: true});
     const {id} = req.params;
+    const maxSize = 6000000;
+    const acceptedExt = ['jpeg', 'jpg', 'png', 'gif']
     
     
     form.parse(req, (err, fields, files) => {
@@ -47,46 +49,48 @@ const modifyArticle = (req, res) => {
             else{ 
                 // si files.files est définit alors on check l'extension
                 const file = files.files  
-                const extensionIsValid =  await checkAcceptedExtensions(file)
+                const extensionIsValid =  await checkAcceptedExtensions(file,acceptedExt)
                 
                 if(!extensionIsValid){
                     // si l'extension n'est pas bonne reponse false
                 res.json({response:false,msg:'Veuillez respecter les formats jpg, jpeg, gif, png'})
+                }
+                else {
+                    if(file.size > maxSize) {
+                    res.json({response: false, msg: 'Le fichier ne doit pas dépasser 6 mo'}) 
                 
             
-                }
-                else{
-                    // si ext. est bonne on remplace cette image par une nouvelle
-         
-                    pool.query(selectPictureUrl,[id], (error, oldUrl, fields) => {
-                        if (error) throw error;
-                        // on selectionne l'url de la photo dans le public/img et la bbd   
-                        const oldUrlPath = 'public/img/'+ oldUrl[0].url;  
-                        const newFilename = files.files.newFilename;
-                        
-                        pool.query(updatePicture, [newFilename, id],(error, newUrl, fields) => {
+                    }
+                    else{
+                        // si ext. est bonne on remplace cette image par une nouvelle
+             
+                        pool.query(selectPictureUrl,[id], (error, oldUrl, fields) => {
                             if (error) throw error;
-                            const newUrlPath = 'public/img/'+ newFilename;  
-                                 
-                            fs.copyFile(files.files.filepath, newUrlPath, (err)=> {
-                                if (err) throw err;
+                            // on selectionne l'url de la photo dans le public/img et la bbd   
+                            const oldUrlPath = 'public/img/'+ oldUrl[0].url;  
+                            const newFilename = files.files.newFilename;
                             
-                                if (newFilename && oldUrl[0].url){
-                                    fs.unlink(oldUrlPath, (err) => {
-                                        if (err) throw err;
-                                        console.log(oldUrl[0].url)
-                                        res.json ({response:true})
-                                    })
-                                } 
+                            pool.query(updatePicture, [newFilename, id],(error, newUrl, fields) => {
+                                if (error) throw error;
+                                const newUrlPath = 'public/img/'+ newFilename;  
+                                     
+                                fs.copyFile(files.files.filepath, newUrlPath, (err)=> {
+                                    if (err) throw err;
+                                
+                                    if (newFilename && oldUrl[0].url){
+                                        fs.unlink(oldUrlPath, (err) => {
+                                            if (err) throw err;
+                                            console.log(oldUrl[0].url)
+                                            res.json ({response:true})
+                                        })
+                                    } 
+                                })
                             })
                         })
-                
-                    })
+                    }
                 }
             }
         })
-
-                    
     })                              
 }
